Damnit, out of space.
Only two fonts in a build :(

No snake :(

But hopefully using less RAM?

And I think everything I actually need need is on it.

Shame we can't just copy fonts from
the flash really.

Hummm.

Would need to keep 4 of them in byte-arrays in
RAM and change them from disk every minute while
the screen was on.

It's a thought.